[DMS-1124] Slice 2 task 2: Extract flattener helpers for resolver reuse
