+++ src/dms/backend/EdFi.DataManagementService.Backend/RelationalWriteMergeSupport.cs
@@ -1,0 +1,162 @@
+// SPDX-License-Identifier: Apache-2.0
+// Licensed to the Ed-Fi Alliance under one or more agreements.
+// The Ed-Fi Alliance licenses this file to you under the Apache License, Version 2.0.
+// See the LICENSE and NOTICES files in the project root for more information.
+
+using System.Collections.Immutable;
+using EdFi.DataManagementService.Backend.External;
+using EdFi.DataManagementService.Backend.External.Plans;
+
+namespace EdFi.DataManagementService.Backend;
+
+/// <summary>
+/// Shared, projection-focused helpers used by the relational write merge
+/// synthesizers (no-profile today, profile-merge in a later slice) and the
+/// post-overlay key-unification resolver. Extracted without behavior change
+/// from <see cref="RelationalWriteNoProfileMergeSynthesizer"/>.
+/// </summary>
+internal static class RelationalWriteMergeSupport
+{
+    internal static int FindBindingIndex(TableWritePlan tableWritePlan, DbColumnName columnName)
+    {
+        for (var bindingIndex = 0; bindingIndex < tableWritePlan.ColumnBindings.Length; bindingIndex++)
+        {
+            if (tableWritePlan.ColumnBindings[bindingIndex].Column.ColumnName.Equals(columnName))
+            {
+                return bindingIndex;
+            }
+        }
+
+        throw new InvalidOperationException(
+            $"Table '{FormatTable(tableWritePlan)}' does not contain a binding for column '{columnName.Value}'."
+        );
+    }
+
+    internal static ImmutableArray<FlattenedWriteValue> ProjectComparableValues(
+        TableWritePlan tableWritePlan,
+        IReadOnlyList<FlattenedWriteValue> values
+    )
+    {
+        var bindingIndexes = tableWritePlan.CollectionMergePlan is null
+            ? Enumerable.Range(0, tableWritePlan.ColumnBindings.Length)
+            : tableWritePlan.CollectionMergePlan.CompareBindingIndexesInOrder;
+
+        FlattenedWriteValue[] comparableValues = bindingIndexes
+            .Select(bindingIndex => values[bindingIndex])
+            .ToArray();
+
+        return comparableValues.ToImmutableArray();
+    }
+
+    internal static object? NormalizeHydratedValue(DbColumnModel column, object? value)
+    {
+        ArgumentNullException.ThrowIfNull(column);
+
+        if (value is null || column.ScalarType is null)
+        {
+            return value;
+        }
+
+        return column.ScalarType.Kind switch
+        {
+            ScalarKind.Date => NormalizeDateValue(value),
+            ScalarKind.Time => NormalizeTimeValue(value),
+            _ => value,
+        };
+    }
+
+    internal static ImmutableArray<RelationalWriteMergedTableRow> ProjectCurrentRows(
+        TableWritePlan tableWritePlan,
+        IReadOnlyList<object?[]> hydratedRows
+    )
+    {
+        List<RelationalWriteMergedTableRow> projectedRows = [];
+
+        foreach (var hydratedRow in hydratedRows)
+        {
+            FlattenedWriteValue[] bindingValues = new FlattenedWriteValue[
+                tableWritePlan.ColumnBindings.Length
+            ];
+
+            for (var bindingIndex = 0; bindingIndex < tableWritePlan.ColumnBindings.Length; bindingIndex++)
+            {
+                var binding = tableWritePlan.ColumnBindings[bindingIndex];
+                var columnOrdinal = FindColumnOrdinal(tableWritePlan.TableModel, binding.Column.ColumnName);
+                bindingValues[bindingIndex] = new FlattenedWriteValue.Literal(
+                    NormalizeHydratedValue(binding.Column, hydratedRow[columnOrdinal])
+                );
+            }
+
+            projectedRows.Add(
+                new RelationalWriteMergedTableRow(
+                    bindingValues,
+                    ProjectComparableValues(tableWritePlan, bindingValues)
+                )
+            );
+        }
+
+        return projectedRows.ToImmutableArray();
+    }
+
+    /// <summary>
+    /// Builds a normalized-by-column projection of a single hydrated table row, keyed
+    /// by column name. Includes every column in <paramref name="rootTableModel"/>.Columns,
+    /// both binding-backed and alias-only. Each value is normalized via
+    /// <see cref="NormalizeHydratedValue"/> so Date/Time comparisons in downstream
+    /// key-unification agreement checks match the binding-indexed projection's
+    /// normalization.
+    /// </summary>
+    internal static IReadOnlyDictionary<DbColumnName, object?> BuildCurrentRootRowByColumnName(
+        DbTableModel rootTableModel,
+        IReadOnlyList<object?> hydratedRow
+    )
+    {
+        ArgumentNullException.ThrowIfNull(rootTableModel);
+        ArgumentNullException.ThrowIfNull(hydratedRow);
+
+        var result = new Dictionary<DbColumnName, object?>();
+        for (var columnOrdinal = 0; columnOrdinal < rootTableModel.Columns.Count; columnOrdinal++)
+        {
+            var column = rootTableModel.Columns[columnOrdinal];
+            result[column.ColumnName] = NormalizeHydratedValue(column, hydratedRow[columnOrdinal]);
+        }
+        return result;
+    }
+
+    private static int FindColumnOrdinal(DbTableModel tableModel, DbColumnName columnName)
+    {
+        for (var columnOrdinal = 0; columnOrdinal < tableModel.Columns.Count; columnOrdinal++)
+        {
+            if (tableModel.Columns[columnOrdinal].ColumnName.Equals(columnName))
+            {
+                return columnOrdinal;
+            }
+        }
+
+        throw new InvalidOperationException(
+            $"Hydrated table '{tableModel.Table.Schema.Value}.{tableModel.Table.Name}' does not contain column '{columnName.Value}'."
+        );
+    }
+
+    private static object NormalizeDateValue(object value) =>
+        value switch
+        {
+            DateOnly => value,
+            DateTime dateTime => DateOnly.FromDateTime(dateTime),
+            DateTimeOffset dateTimeOffset => DateOnly.FromDateTime(dateTimeOffset.DateTime),
+            _ => value,
+        };
+
+    private static object NormalizeTimeValue(object value) =>
+        value switch
+        {
+            TimeOnly => value,
+            TimeSpan timeSpan => TimeOnly.FromTimeSpan(timeSpan),
+            DateTime dateTime => TimeOnly.FromDateTime(dateTime),
+            DateTimeOffset dateTimeOffset => TimeOnly.FromDateTime(dateTimeOffset.DateTime),
+            _ => value,
+        };
+
+    private static string FormatTable(TableWritePlan tableWritePlan) =>
+        $"{tableWritePlan.TableModel.Table.Schema.Value}.{tableWritePlan.TableModel.Table.Name}";
+}
